requirements.txt
allure-pytest
faker
jsonpath-rw-ext
nose
pytest
python-dateutil
requests
selenium